pad changes, title bar, game over screen
a lot of stuff was written through the standard screen instead of the pad. This was fixed, clearing the way for better, uh, paddification (ie levels larger than the window) later.
this also cleaned a lot of stuff, since now we can use the coordinates themselves and not xCoord-1 or whatever

additionally, I added a title bar : it displays the name of the level and the number of moves.

I also changed the game over screen a bit : it now displays the number of moves and offers to quit