infinite levels and resizing
through the use of pads, we now support scrolling the level as we go ! This means a theoritically infinite max level size.
Also, it does mean you can resize the window much more easily.
In fact, the big constraint now is the title screen, which much be at least ~30 chars wide ; in game, the program stays viable (ie you can play it) down to 5x5 sizes !